move /user/info test suite to user.test.js, fix race condition
